(PC-7698) Define private_api and public_api in a separate module
Previsouly, routes modules had to import `pcapi.flask_app`, which
caused import loops.
