+++ src/pcapi/routes/internal/health_check.py
@@ -1,4 +1,4 @@
-from pcapi.flask_app import public_api
+from pcapi.routes.apis import public_api
 from pcapi.utils.health_checker import check_database_connection
 from pcapi.utils.health_checker import read_version_from_file
