+++ src/pcapi/routes/pro/signup.py
@@ -1,5 +1,5 @@
 from pcapi.core.users.api import create_pro_user_and_offerer
-from pcapi.flask_app import private_api
+from pcapi.routes.apis import private_api
 from pcapi.routes.serialization.users import ProUserCreationBodyModel
 from pcapi.serialization.decorator import spectree_serialize
