+++ src/pcapi/routes/pro/users.py
@@ -4,8 +4,8 @@ from flask_login import login_required
 
 from pcapi.core.users import api as users_api
 from pcapi.core.users.models import User
-from pcapi.flask_app import private_api
 from pcapi.models import ApiErrors
+from pcapi.routes.apis import private_api
 from pcapi.routes.serialization.users import PatchProUserBodyModel
 from pcapi.routes.serialization.users import PatchProUserResponseModel
 from pcapi.serialization.decorator import spectree_serialize
