+++ src/pcapi/routes/pro/venue_providers.py
@@ -10,8 +10,8 @@ from pcapi.core.providers.api import update_allocine_venue_provider
 from pcapi.core.providers.models import AllocineVenueProvider
 from pcapi.core.providers.models import VenueProviderCreationPayload
 from pcapi.core.providers.repository import get_venue_provider_by_venue_and_provider_ids
-from pcapi.flask_app import private_api
 from pcapi.models import ApiErrors
+from pcapi.routes.apis import private_api
 from pcapi.routes.serialization.venue_provider_serialize import ListVenueProviderQuery
 from pcapi.routes.serialization.venue_provider_serialize import ListVenueProviderResponse
 from pcapi.routes.serialization.venue_provider_serialize import PostVenueProviderBody
