+++ src/pcapi/routes/shared/features.py
@@ -1,7 +1,7 @@
 from flask import jsonify
 
-from pcapi.flask_app import public_api
 from pcapi.repository import feature_queries
+from pcapi.routes.apis import public_api
 from pcapi.routes.serialization import as_dict
 from pcapi.utils.includes import FEATURE_INCLUDES
