+++ src/pcapi/routes/webapp/mailing_contacts.py
@@ -1,6 +1,6 @@
 from flask import request
 
-from pcapi.flask_app import public_api
+from pcapi.routes.apis import public_api
 from pcapi.utils.rest import expect_json_data
 from pcapi.validation.routes.mailing_contacts import validate_save_mailing_contact_request
 from pcapi.workers.mailing_contacts_job import mailing_contacts_job
