+++ src/pcapi/routes/webapp/music_types.py
@@ -2,7 +2,7 @@ from flask import jsonify
 from flask_login import login_required
 
 from pcapi.domain.music_types import music_types
-from pcapi.flask_app import private_api
+from pcapi.routes.apis import private_api
 
 
 @private_api.route("/musicTypes", methods=["GET"])
