+++ src/pcapi/routes/webapp/show_types.py
@@ -2,7 +2,7 @@ from flask import jsonify
 from flask_login import login_required
 
 from pcapi.domain.show_types import show_types
-from pcapi.flask_app import private_api
+from pcapi.routes.apis import private_api
 
 
 @private_api.route("/showTypes", methods=["GET"])
